# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "블로그 백엔드 API 서버",
        "title": "Bumsiku API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/comments/{commentId}": {
            "delete": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "블로그 댓글을 삭제합니다 (관리자 전용)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "댓글"
                ],
                "summary": "댓글 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "댓글 ID",
                        "name": "commentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제 성공 메시지",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "인증 실패",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "댓글을 찾을 수 없음",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/posts": {
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "새 블로그 게시물을 작성합니다 (관리자 전용)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "게시물"
                ],
                "summary": "게시물 작성",
                "parameters": [
                    {
                        "description": "게시물 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreatePostRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Post"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "인증 실패",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/posts/{id}": {
            "put": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "기존 블로그 게시물을 수정합니다 (관리자 전용)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "게시물"
                ],
                "summary": "게시물 수정",
                "parameters": [
                    {
                        "type": "string",
                        "description": "게시물 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "수정할 게시물 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdatePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Post"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "인증 실패",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "게시물을 찾을 수 없음",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "블로그 게시물과 관련 댓글을 삭제합니다 (관리자 전용)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "게시물"
                ],
                "summary": "게시물 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "게시물 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제 성공 메시지",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "인증 실패",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "게시물을 찾을 수 없음",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/comments/{id}": {
            "get": {
                "description": "특정 게시물에 작성된 댓글 목록을 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "댓글"
                ],
                "summary": "게시물 댓글 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "게시물 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "댓글 목록",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/comments/{postId}": {
            "post": {
                "description": "특정 게시물에 새 댓글을 등록합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "댓글"
                ],
                "summary": "댓글 등록",
                "parameters": [
                    {
                        "type": "string",
                        "description": "게시물 ID",
                        "name": "postId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "댓글 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Comment"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "게시물을 찾을 수 없음",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "블로그 관리자 로그인 API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "인증"
                ],
                "summary": "관리자 로그인",
                "parameters": [
                    {
                        "description": "로그인 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "로그인 성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "로그인 실패",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "description": "블로그 게시물 목록을 페이지네이션하여 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "게시물"
                ],
                "summary": "게시물 목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "카테고리 필터",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "페이지 번호 (기본값: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "페이지 크기 (기본값: 10)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetPostsResponse"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "description": "특정 ID의 게시물 상세 정보를 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "게시물"
                ],
                "summary": "게시물 상세 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "게시물 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Post"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "게시물을 찾을 수 없음",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "오류 코드",
                    "type": "string",
                    "example": "BAD_REQUEST"
                },
                "message": {
                    "description": "오류 메시지",
                    "type": "string",
                    "example": "잘못된 요청입니다"
                }
            }
        },
        "handler.CreateCommentRequest": {
            "type": "object",
            "required": [
                "content",
                "nickname"
            ],
            "properties": {
                "content": {
                    "description": "댓글 내용",
                    "type": "string",
                    "example": "좋은 글이네요!"
                },
                "nickname": {
                    "description": "닉네임",
                    "type": "string",
                    "example": "익명사용자"
                }
            }
        },
        "handler.CreatePostRequest": {
            "type": "object",
            "required": [
                "category",
                "content",
                "summary",
                "title"
            ],
            "properties": {
                "category": {
                    "description": "카테고리",
                    "type": "string",
                    "example": "technology"
                },
                "content": {
                    "description": "게시물 내용",
                    "type": "string",
                    "example": "게시물 본문 내용..."
                },
                "summary": {
                    "description": "게시물 요약",
                    "type": "string",
                    "example": "게시물 요약..."
                },
                "title": {
                    "description": "게시물 제목",
                    "type": "string",
                    "example": "새로운 블로그 게시물"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "오류 정보",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handler.APIError"
                        }
                    ]
                },
                "success": {
                    "description": "요청 성공 여부 (항상 false)",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "handler.GetPostsResponse": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "description": "현재 페이지",
                    "type": "integer",
                    "example": 1
                },
                "posts": {
                    "description": "게시물 목록",
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "totalCount": {
                    "description": "전체 게시물 수",
                    "type": "integer",
                    "example": 100
                },
                "totalPages": {
                    "description": "전체 페이지 수",
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "handler.UpdatePostRequest": {
            "type": "object",
            "required": [
                "category",
                "content",
                "summary",
                "title"
            ],
            "properties": {
                "category": {
                    "description": "카테고리",
                    "type": "string",
                    "example": "technology"
                },
                "content": {
                    "description": "게시물 내용",
                    "type": "string",
                    "example": "수정된 게시물 본문 내용..."
                },
                "summary": {
                    "description": "게시물 요약",
                    "type": "string",
                    "example": "수정된 게시물 요약..."
                },
                "title": {
                    "description": "게시물 제목",
                    "type": "string",
                    "example": "수정된 블로그 게시물"
                }
            }
        },
        "model.Comment": {
            "type": "object",
            "properties": {
                "commentId": {
                    "description": "댓글 ID",
                    "type": "string",
                    "example": "comment-123"
                },
                "content": {
                    "description": "댓글 내용",
                    "type": "string",
                    "example": "댓글 내용입니다."
                },
                "createdAt": {
                    "description": "생성 시간",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "nickname": {
                    "description": "닉네임",
                    "type": "string",
                    "example": "익명사용자"
                },
                "postId": {
                    "description": "게시물 ID",
                    "type": "string",
                    "example": "post-123"
                }
            }
        },
        "model.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "비밀번호",
                    "type": "string",
                    "example": "********"
                },
                "username": {
                    "description": "사용자 아이디",
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "model.Post": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "카테고리",
                    "type": "string",
                    "example": "technology"
                },
                "content": {
                    "description": "게시물 내용",
                    "type": "string",
                    "example": "게시물 본문 내용..."
                },
                "createdAt": {
                    "description": "생성 시간",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "postId": {
                    "description": "게시물 ID",
                    "type": "string",
                    "example": "post-123"
                },
                "summary": {
                    "description": "게시물 요약",
                    "type": "string",
                    "example": "게시물 요약..."
                },
                "title": {
                    "description": "게시물 제목",
                    "type": "string",
                    "example": "블로그 제목"
                },
                "updatedAt": {
                    "description": "수정 시간",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "관리자 인증 세션 쿠키",
            "type": "apiKey",
            "name": "loginSession",
            "in": "cookie"
        }
    }
}